% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{Annotator.annotate}
\alias{Annotator.annotate}
\title{Annotator.annotate}
\usage{
Annotator.annotate(
  records,
  fields,
  include_errors = FALSE,
  raw = FALSE,
  env = solvebio:::.solveEnv
)
}
\arguments{
\item{records}{The data table or data frame to annotate.}

\item{fields}{The fields to add.}

\item{include_errors}{Set to TRUE to include errors in the output (default: FALSE).}

\item{raw}{Set to TRUE to return the raw response (default: FALSE).}

\item{env}{(optional) Custom client environment.}
}
\description{
Annotate a data table/frame with additional fields.
}
\examples{
\dontrun{
Annotator.annotate(records=tbl, fields=fields)
}

}
\references{
\url{https://docs.solvebio.com/}
}
