% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\name{Application.all}
\alias{Application.all}
\title{Application.all}
\usage{
Application.all(env = .solveEnv, ...)
}
\arguments{
\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters.}
}
\description{
Retrieves the metadata about all application on SolveBio available to the current user.
}
\examples{
\dontrun{
Application.all()
}

}
\references{
\url{https://docs.solvebio.com/}
}
