% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\name{Application.create}
\alias{Application.create}
\title{Application.create}
\usage{
Application.create(name, description, env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{name}{The name of the application.}

\item{description}{(optional) An optional description for the application.}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional application attributes.}
}
\description{
Create a new SolveBio application.
}
\examples{
\dontrun{
Application.create(
                   name="My new application",
                   )
}

}
\references{
\url{https://docs.solvebio.com/}
}
