% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\name{Application.retrieve}
\alias{Application.retrieve}
\title{Application.retrieve}
\usage{
Application.retrieve(client_id, env = .solveEnv)
}
\arguments{
\item{client_id}{The client ID for the application.}

\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the metadata about a specific application SolveBio.
}
\examples{
\dontrun{
Application.retrieve("abcd1234")
}

}
\references{
\url{https://docs.solvebio.com/}
}
