% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\name{Application.shinyServer}
\alias{Application.shinyServer}
\title{Application.shinyServer}
\usage{
Application.shinyServer(client_id, server, base_url)
}
\arguments{
\item{client_id}{The application's client ID.}

\item{server}{The Shiny server function.}

\item{base_url}{(optional) Override the default login host (default: https://my.solvebio.com).}
}
\description{
Wraps an existing Shiny server in an OAuth2 flow.
}
\examples{
\dontrun{
Application.shinyServer(
                        client_id="abcd1234",
                        server=server,
                       )
}

}
\references{
\url{https://docs.solvebio.com/}
}
