% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\name{Application.update}
\alias{Application.update}
\title{Application.update}
\usage{
Application.update(client_id, env = .solveEnv, ...)
}
\arguments{
\item{client_id}{The client ID for the application.}

\item{env}{(optional) Custom client environment.}

\item{...}{Application attributes to change.}
}
\description{
Updates the attributes of an existing application.
}
\examples{
\dontrun{
Application.update(
                   "abcd1234",
                   name="New app name"
                   )
}

}
\references{
\url{https://docs.solvebio.com/}
}
