% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beacon.R
\name{Beacon.delete}
\alias{Beacon.delete}
\title{Beacon.delete}
\usage{
Beacon.delete(id, env = .solveEnv)
}
\arguments{
\item{id}{The ID of the beacon.}

\item{env}{(optional) Custom client environment.}
}
\description{
Delete a specific beacon from SolveBio.
}
\examples{
\dontrun{
Beacon.delete("1234567890")
}

}
\references{
\url{https://docs.solvebio.com/}
}
