% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beacon.R
\name{Beacon.query}
\alias{Beacon.query}
\title{Beacon.query}
\usage{
Beacon.query(id, query, entity_type, env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{id}{The ID of the beacon.}

\item{query}{The entity ID or query string.}

\item{entity_type}{(optional) A valid SolveBio entity type.}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters.}
}
\description{
Query an individual beacon.
}
\examples{
\dontrun{
Beacon.query(
             id="1234",
             query="BRCA2",
             entity_type="gene"
             )
}

}
\references{
\url{https://docs.solvebio.com/}
}
