% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beacon.R
\name{Beacon.retrieve}
\alias{Beacon.retrieve}
\title{Beacon.retrieve}
\usage{
Beacon.retrieve(id)
}
\arguments{
\item{id}{The ID of the beacon.}
}
\description{
Retrieves the metadata about a specific beacon on SolveBio.
}
\examples{
\dontrun{
Beacon.retrieve("1234")
}

}
\references{
\url{https://docs.solvebio.com/}
}
