% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beacon.R
\name{Beacon.update}
\alias{Beacon.update}
\title{Beacon.update}
\usage{
Beacon.update(id, env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{id}{The ID of the beacon to update.}

\item{env}{(optional) Custom client environment.}

\item{...}{Beacon attributes to change.}
}
\description{
Updates the attributes of an existing beacon.
}
\examples{
\dontrun{
Beacon.update(
              id="1234",
              title="New Beacon Title"
             )
}

}
\references{
\url{https://docs.solvebio.com/}
}
