% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beacon_set.R
\name{BeaconSet.all}
\alias{BeaconSet.all}
\title{BeaconSet.all}
\usage{
BeaconSet.all(env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters.}
}
\description{
Retrieves the metadata about all beacon sets on SolveBio accessible to the current user.
}
\examples{
\dontrun{
BeaconSet.all()
}

}
\references{
\url{https://docs.solvebio.com/}
}
