% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beacon_set.R
\name{BeaconSet.create}
\alias{BeaconSet.create}
\title{BeaconSet.create}
\usage{
BeaconSet.create(
  title,
  description,
  is_shared = FALSE,
  env = solvebio:::.solveEnv,
  ...
)
}
\arguments{
\item{title}{The title displayed for the beacon set.}

\item{description}{(optional) An optional description for the new beacon set.}

\item{is_shared}{If TRUE, everyone else in your account will be able to see and query the beacon set, but will not be able to edit it. (Default: FALSE)}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional beacon set attributes.}
}
\description{
Create a new beacon set.
}
\examples{
\dontrun{
BeaconSet.create(
                 title="My new beacon set",
                 )
}

}
\references{
\url{https://docs.solvebio.com/}
}
