% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beacon_set.R
\name{BeaconSet.query}
\alias{BeaconSet.query}
\title{BeaconSet.query}
\usage{
BeaconSet.query(id, query, entity_type, env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{id}{The ID of the beacon set.}

\item{query}{The entity ID or query string.}

\item{entity_type}{(optional) A valid SolveBio entity type.}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters.}
}
\description{
Query a beacon set (i.e. all the beacons within a beacon set).
}
\examples{
\dontrun{
BeaconSet.query(
                id="1234",
                query="BRCA2",
                entity_type="gene"
                )
}

}
\references{
\url{https://docs.solvebio.com/}
}
