% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvebio.R
\name{createEnv}
\alias{createEnv}
\title{createEnv}
\usage{
createEnv(token, token_type = "Token", host = "https://api.solvebio.com")
}
\arguments{
\item{token}{A SolveBio API key or OAuth2 token}

\item{token_type}{SolveBio token type (default: Token)}

\item{host}{(optional) The SolveBio API host (default: https://api.solvebio.com)}
}
\description{
Create a new SolveBio environment.
}
\examples{
\dontrun{
env <- createEnv("MyAPIkey")
User.retrieve(env = myEnv)
}

}
\references{
\url{https://docs.solvebio.com/}
}
