% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset.activity}
\alias{Dataset.activity}
\title{Dataset.activity}
\usage{
Dataset.activity(id, follow = TRUE, env = .solveEnv)
}
\arguments{
\item{id}{String The ID of a SolveBio dataset}

\item{follow}{Follow active tasks until they complete.}

\item{env}{(optional) Custom client environment.}
}
\description{
A helper function to get or follow the current activity on a dataset.
}
\examples{
\dontrun{
Dataset.activity("1234567890")
}

}
\references{
\url{https://docs.solvebio.com/}
}
