% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/object.R
\name{Dataset.all}
\alias{Dataset.all}
\alias{Dataset.all}
\title{Dataset.all}
\usage{
Dataset.all(...)

Dataset.all(...)
}
\arguments{
\item{...}{(optional) Additional query parameters (e.g. page).}

\item{...}{(optional) Additional query parameters (e.g. page).}
}
\description{
Retrieves the metadata about all datasets on SolveBio.

Retrieves the metadata about all datasets on SolveBio.
}
\examples{
\dontrun{
Dataset.all()
}

\dontrun{
Dataset.all()
}

}
\references{
\url{https://docs.solvebio.com/}

\url{https://docs.solvebio.com/}
}
