% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset.create}
\alias{Dataset.create}
\title{Dataset.create}
\usage{
Dataset.create(depository_version_id, name, ...)
}
\arguments{
\item{depository_version_id}{The ID of the parent depository verison.}

\item{name}{The unique name of the dataset witin the version.}

\item{...}{(optional) Additional dataset attributes.}
}
\description{
Create an empty SolveBio dataset.
}
\examples{
\dontrun{
Dataset.create(depository_version_id=1, name="MyDataset")
}

}
\references{
\url{https://docs.solvebio.com/}
}
