% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset.data}
\alias{Dataset.data}
\title{Dataset.data}
\usage{
Dataset.data(id, filters, ...)
}
\arguments{
\item{id}{The ID or full name of a SolveBio dataset, or a Dataset object.}

\item{filters}{(optional) Query filters.}

\item{...}{(optional) Additional query parameters (e.g. limit, offset).}
}
\description{
Returns one page of documents from a SolveBio dataset and processes the response.
}
\examples{
\dontrun{
login()
Dataset.data("ClinVar/ClinVar")
}

}
\references{
\url{https://docs.solvebio.com/}
}

