% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/object.R
\name{Dataset.delete}
\alias{Dataset.delete}
\alias{Dataset.delete}
\title{Dataset.delete}
\usage{
Dataset.delete(id)

Dataset.delete(id)
}
\arguments{
\item{id}{String The ID or full name of a SolveBio dataset}

\item{id}{String The ID or full name of a SolveBio dataset}
}
\description{
Delete a specific dataset from SolveBio.

Delete a specific dataset from SolveBio.
}
\examples{
\dontrun{
Dataset.delete("1")
}

\dontrun{
Dataset.delete("1")
}

}
\references{
\url{https://docs.solvebio.com/}

\url{https://docs.solvebio.com/}
}
