% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset.enable_global_beacon}
\alias{Dataset.enable_global_beacon}
\title{Dataset.enable_global_beacon}
\usage{
Dataset.enable_global_beacon(id, env = solvebio:::.solveEnv)
}
\arguments{
\item{id}{The ID of a SolveBio dataset.}

\item{env}{(optional) Custom client environment.}
}
\description{
Enables Global Beacon for the the dataset.
}
\examples{
\dontrun{
Dataset.enable_global_beacon("1234567890")
}

}
\references{
\url{https://docs.solvebio.com/}
}
