% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset.get_by_full_path}
\alias{Dataset.get_by_full_path}
\title{Dataset.get_by_full_path}
\usage{
Dataset.get_by_full_path(full_path, env = .solveEnv)
}
\arguments{
\item{full_path}{A valid full path to a dataset.}

\item{env}{(optional) Custom client environment.}
}
\description{
A helper function to get a dataset by its full path.
}
\examples{
\dontrun{
Dataset.get_by_full_path("solvebio:public:/ClinVar/3.7.4-2017-01-30/Variants-GRCh37")
}

}
\references{
\url{https://docs.solvebio.com/}
}
