% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset.get_or_create_by_full_name}
\alias{Dataset.get_or_create_by_full_name}
\title{Dataset.get_or_create_by_full_name}
\usage{
Dataset.get_or_create_by_full_name(full_name, ...)
}
\arguments{
\item{full_name}{A valid dataset full name (<depository/version/dataset>).}

\item{...}{(optional) Additional dataset attributes.}
}
\description{
A helper function to create a dataset on SolveBio using a full name.
}
\examples{
\dontrun{
Dataset.get_or_create_by_full_name(<FULL NAME>)
}

}
\references{
\url{https://docs.solvebio.com/}
}

