% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset.get_or_create_by_full_path}
\alias{Dataset.get_or_create_by_full_path}
\title{Dataset.get_or_create_by_full_path}
\usage{
Dataset.get_or_create_by_full_path(full_path, ...)
}
\arguments{
\item{full_path}{A valid full path to a dataset.}

\item{...}{Additional dataset creation parameters.}
}
\description{
A helper function to get or create a dataset by its full path.
}
\examples{
\dontrun{
Dataset.get_or_create_by_full_path("MyVault:/folder/sub-folder/dataset")
}

}
\references{
\url{https://docs.solvebio.com/}
}
