% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset.query}
\alias{Dataset.query}
\title{Dataset.query}
\usage{
Dataset.query(id, ...)
}
\arguments{
\item{id}{String The ID or full name of a SolveBio dataset, or a Dataset object.}

\item{filters}{Json (optional) Query filters.}

\item{params}{Json (optional) Query parameters.}
}
\description{
Returns filtered documents from a SolveBio dataset.
\code{query} executes a SolveBio dataset query and retrieves the results.
}
\examples{
\dontrun{
login()
Dataset.query("ClinVar/ClinVar")
}
}
\references{
\url{https://docs.solvebio.com/}
}

