% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/object.R
\name{Dataset.retrieve}
\alias{Dataset.retrieve}
\alias{Dataset.retrieve}
\title{Dataset.retrieve}
\usage{
Dataset.retrieve(id)

Dataset.retrieve(id)
}
\arguments{
\item{id}{String The ID or full name of a SolveBio dataset}

\item{id}{String The ID or full name of a SolveBio dataset}
}
\description{
Retrieves the metadata about a specific dataset from SolveBio.

Retrieves the metadata about a specific dataset from SolveBio.
}
\examples{
\dontrun{
Dataset.retrieve("ClinVar/ClinVar")
}

\dontrun{
Dataset.retrieve("ClinVar/ClinVar")
}

}
\references{
\url{https://docs.solvebio.com/}

\url{https://docs.solvebio.com/}
}
