% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset.update}
\alias{Dataset.update}
\title{Dataset.update}
\usage{
Dataset.update(id, ...)
}
\arguments{
\item{id}{The ID of the dataset to update.}

\item{...}{Dataset attributes to change.}
}
\description{
Updates the attributes of an existing dataset.
}
\examples{
\dontrun{
Dataset.update(
               id="1234",
               name="New Dataset Name",
              )
}

}
\references{
\url{https://docs.solvebio.com/}
}
