% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_commit.R
\name{DatasetCommit.delete}
\alias{DatasetCommit.delete}
\title{DatasetCommit.delete}
\usage{
DatasetCommit.delete(id, env = .solveEnv)
}
\arguments{
\item{id}{String The ID or full name of a SolveBio dataset commit.}

\item{env}{(optional) Custom client environment.}
}
\description{
Deletes a specific dataset commit on SolveBio.
}
\examples{
\dontrun{
DatasetCommit.delete(<ID>)
}

}
\references{
\url{https://docs.solvebio.com/}
}
