% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_commit.R
\name{DatasetCommit.retrieve}
\alias{DatasetCommit.retrieve}
\title{DatasetCommit.retrieve}
\usage{
DatasetCommit.retrieve(id)
}
\arguments{
\item{id}{String The ID or full name of a SolveBio dataset commit.}
}
\description{
Retrieves the metadata about a specific dataset commit on SolveBio.
}
\examples{
\dontrun{
DatasetCommit.retrieve(<ID>)
}

}
\references{
\url{https://docs.solvebio.com/}
}
