% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_export.R
\name{DatasetExport.create}
\alias{DatasetExport.create}
\title{DatasetExport.create}
\usage{
DatasetExport.create(dataset_id, format = "json", params = list(), ...)
}
\arguments{
\item{dataset_id}{The target dataset ID.}

\item{format}{(optional) The export format (default: json).}

\item{params}{(optional) Query parameters for the export.}

\item{...}{(optional) Additional dataset export parameters.}
}
\description{
Create a new dataset export.
}
\examples{
\dontrun{
DatasetExport.create(dataset_id=<ID>, format='json', params=list(fields=c("field_1"), limit=100))
}

}
\references{
\url{https://docs.solvebio.com/}
}
