% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_field.R
\name{DatasetField.facets}
\alias{DatasetField.facets}
\title{DatasetField.facets}
\usage{
DatasetField.facets(id, ...)
}
\arguments{
\item{id}{String The ID or full name of a dataset field.}

\item{...}{(optional) Additional query parameters.}
}
\description{
Returns the facets for a given dataset field.
}
\examples{
\dontrun{
DatasetField.facets(691)
}

}
\references{
\url{https://docs.solvebio.com/}
}

