% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_field.R
\name{DatasetField.update}
\alias{DatasetField.update}
\title{DatasetField.update}
\usage{
DatasetField.update(id, env = .solveEnv, ...)
}
\arguments{
\item{id}{The ID of the dataset field to update.}

\item{env}{(optional) Custom client environment.}

\item{...}{Dataset field attributes to change.}
}
\description{
Updates the attributes of an existing dataset field. NOTE: The data_type of a field cannot be changed.
}
\examples{
\dontrun{
DatasetField.update(
                    id="1234",
                    title="New Field Title"
                   )
}

}
\references{
\url{https://docs.solvebio.com/}
}
