% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_import.R
\name{DatasetImport.delete}
\alias{DatasetImport.delete}
\title{DatasetImport.delete}
\usage{
DatasetImport.delete(id)
}
\arguments{
\item{id}{String The ID of a SolveBio dataset import.}
}
\description{
Deletes a specific dataset import on SolveBio.
}
\examples{
\dontrun{
DatasetImport.delete(<ID>)
}

}
\references{
\url{https://docs.solvebio.com/}
}
