% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_import.R
\name{DatasetImport.retrieve}
\alias{DatasetImport.retrieve}
\title{DatasetImport.retrieve}
\usage{
DatasetImport.retrieve(id)
}
\arguments{
\item{id}{String The ID or full name of a SolveBio dataset import.}
}
\description{
Retrieves the metadata about a specific dataset import on SolveBio.
}
\examples{
\dontrun{
DatasetImport.retrieve(<ID>)
}

}
\references{
\url{https://docs.solvebio.com/}
}

