% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_migration.R
\name{DatasetMigration.create}
\alias{DatasetMigration.create}
\title{DatasetMigration.create}
\usage{
DatasetMigration.create(source_id, target_id, commit_mode = "append",
  source_params = NULL, target_fields = NULL, include_errors = FALSE, ...)
}
\arguments{
\item{source_id}{The source dataset ID.}

\item{target_id}{The target dataset ID.}

\item{commit_mode}{(optional) The commit mode (default: append).}

\item{source_params}{(optional) The query parameters used on the source dataset.}

\item{target_fields}{(optional) A list of valid dataset fields to add or override in the target dataset.}

\item{include_errors}{(optional) If TRUE, a new field (_errors) will be added to each record containing expression evaluation errors (default: FALSE).}

\item{...}{(optional) Additional dataset migration attributes.}
}
\description{
Create a new dataset migration.
}
\examples{
\dontrun{
DatasetMigration.create(dataset_id=<ID>, upload_id=<ID>)
}

}
\references{
\url{https://docs.solvebio.com/}
}
