% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_migration.R
\name{DatasetMigration.delete}
\alias{DatasetMigration.delete}
\title{DatasetMigration.delete}
\usage{
DatasetMigration.delete(id, env = .solveEnv)
}
\arguments{
\item{id}{String The ID of a SolveBio dataset migration.}

\item{env}{(optional) Custom client environment.}
}
\description{
Deletes specific dataset migration on SolveBio.
}
\examples{
\dontrun{
DatasetMigration.delete(<ID>)
}

}
\references{
\url{https://docs.solvebio.com/}
}
