% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_migration.R
\name{DatasetMigration.retrieve}
\alias{DatasetMigration.retrieve}
\title{DatasetMigration.retrieve}
\usage{
DatasetMigration.retrieve(id, env = .solveEnv)
}
\arguments{
\item{id}{String The ID of a SolveBio dataset migration.}

\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the metadata about a specific dataset migration on SolveBio.
}
\examples{
\dontrun{
DatasetMigration.retrieve(<ID>)
}

}
\references{
\url{https://docs.solvebio.com/}
}
