% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_template.R
\name{DatasetTemplate.all}
\alias{DatasetTemplate.all}
\title{DatasetTemplate.all}
\usage{
DatasetTemplate.all(env = .solveEnv, ...)
}
\arguments{
\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters (e.g. page).}
}
\description{
Retrieves the metadata about all dataset templates on SolveBio.
}
\examples{
\dontrun{
DatasetTemplate.all()
}

}
\references{
\url{https://docs.solvebio.com/}
}
