% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_template.R
\name{DatasetTemplate.create}
\alias{DatasetTemplate.create}
\title{DatasetTemplate.create}
\usage{
DatasetTemplate.create(env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Dataset template attributes.}
}
\description{
Create a SolveBio dataset template.
}
\examples{
\dontrun{
DatasetTemplate.create(name="My Dataset Template")
}

}
\references{
\url{https://docs.solvebio.com/}
}
