% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_template.R
\name{DatasetTemplate.delete}
\alias{DatasetTemplate.delete}
\title{DatasetTemplate.delete}
\usage{
DatasetTemplate.delete(id, env = solvebio:::.solveEnv)
}
\arguments{
\item{id}{String The ID or full name of a SolveBio dataset template.}

\item{env}{(optional) Custom client environment.}
}
\description{
Deletes a specific dataset template on SolveBio.
}
\examples{
\dontrun{
DatasetTemplate.delete(<ID>)
}

}
\references{
\url{https://docs.solvebio.com/}
}
