% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_template.R
\name{DatasetTemplate.retrieve}
\alias{DatasetTemplate.retrieve}
\title{DatasetTemplate.retrieve}
\usage{
DatasetTemplate.retrieve(id, env = solvebio:::.solveEnv)
}
\arguments{
\item{id}{String The ID of a SolveBio dataset template.}

\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the metadata about a specific dataset template on SolveBio.
}
\examples{
\dontrun{
DatasetTemplate.retrieve(<ID>)
}

}
\references{
\url{https://docs.solvebio.com/}
}
