% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_template.R
\name{DatasetTemplate.update}
\alias{DatasetTemplate.update}
\title{DatasetTemplate.update}
\usage{
DatasetTemplate.update(id, env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{id}{The ID of the dataset template to update.}

\item{env}{(optional) Custom client environment.}

\item{...}{Dataset template attributes to change.}
}
\description{
Updates the attributes of an existing dataset template.
}
\examples{
\dontrun{
DatasetTemplate.update(
               id="1234",
               name="New Template Name",
              )
}

}
\references{
\url{https://docs.solvebio.com/}
}
