% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depository.R
\name{Depository.all}
\alias{Depository.all}
\title{Depository.all}
\usage{
Depository.all(...)
}
\arguments{
\item{...}{(optional) Additional query parameters (e.g. limit, offset).}
}
\description{
Retrieves the metadata about all depositories on SolveBio.
}
\examples{
\dontrun{
Depository.all()
}

}
\references{
\url{https://docs.solvebio.com/}
}
