% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depository.R
\name{Depository.create}
\alias{Depository.create}
\title{Depository.create}
\usage{
Depository.create(name, ...)
}
\arguments{
\item{name}{The unique name of the depository.}

\item{...}{(optional) Additional depository attributes.}
}
\description{
Create a new SolveBio depository.
}
\examples{
\dontrun{
Depository.create(name="my-domain:MyDepository")
}

}
\references{
\url{https://docs.solvebio.com/}
}
