% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depository.R
\name{Depository.delete}
\alias{Depository.delete}
\title{Depository.delete}
\usage{
Depository.delete(id)
}
\arguments{
\item{id}{String The ID or full name of a SolveBio depository.}
}
\description{
Delete a specific depository from SolveBio.
}
\examples{
\dontrun{
Depository.delete("1")
}

}
\references{
\url{https://docs.solvebio.com/}
}
