% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depository.R
\name{Depository.latest_version}
\alias{Depository.latest_version}
\title{Depository.latest_version}
\usage{
Depository.latest_version(id, ...)
}
\arguments{
\item{id}{String The ID or full name of a depository, or a Depository object.}

\item{...}{(optional) Additional query parameters.}
}
\description{
Retrieves the latest version for a given depository.
}
\examples{
\dontrun{
Depository.latest_version("ClinVar")
}

}
\references{
\url{https://docs.solvebio.com/}
}
