% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depository.R
\name{Depository.retrieve}
\alias{Depository.retrieve}
\title{Depository.retrieve}
\usage{
Depository.retrieve(id)
}
\arguments{
\item{id}{String The ID or full name of a SolveBio depository}
}
\description{
Retrieves the metadata about a specific SolveBio depository.
}
\examples{
\dontrun{
Depository.retrieve("ClinVar")
}

}
\references{
\url{https://docs.solvebio.com/}
}
