% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/depository.R
\name{Depository.versions}
\alias{Depository.versions}
\title{Depository.versions}
\usage{
Depository.versions(id, ...)
}
\arguments{
\item{id}{String The ID or full name of a SolveBio depository.}
}
\description{
Returns the list of versions for a given depository.
}
\examples{
\dontrun{
Depository.versions("ClinVar")
}
}
\references{
\url{https://docs.solvebio.com/}
}

