% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/depository_version.R
\name{DepositoryVersion.all}
\alias{DepositoryVersion.all}
\title{DepositoryVersion.all}
\usage{
DepositoryVersion.all(...)
}
\description{
Retrieves the metadata about all depository versions on SolveBio.
}
\examples{
\dontrun{
DepositoryVersion.all()
}
}
\references{
\url{https://docs.solvebio.com/}
}

