% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depository_version.R
\name{DepositoryVersion.create}
\alias{DepositoryVersion.create}
\title{DepositoryVersion.create}
\usage{
DepositoryVersion.create(depository_id, name, ...)
}
\arguments{
\item{depository_id}{The ID of the parent depository.}

\item{name}{The semantic version (i.e. "1.0.0").}

\item{...}{(optional) Additional depository version attributes.}
}
\description{
Create a new SolveBio depository version, within a depository.
}
\examples{
\dontrun{
DepositoryVersion.create(depository_id=<DEPOSITORY ID>, name="1.0.0")
}

}
\references{
\url{https://docs.solvebio.com/}
}
