% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depository_version.R
\name{DepositoryVersion.datasets}
\alias{DepositoryVersion.datasets}
\title{DepositoryVersion.datasets}
\usage{
DepositoryVersion.datasets(id, ...)
}
\arguments{
\item{id}{String The ID or full name of a depository version.}

\item{...}{(optional) Additional query parameters (e.g. page).}
}
\description{
Returns the list of datasets for a given depository version.
}
\examples{
\dontrun{
DepositoryVersion.datasets("ClinVar/1.0.0")
}

}
\references{
\url{https://docs.solvebio.com/}
}
