% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depository_version.R
\name{DepositoryVersion.retrieve}
\alias{DepositoryVersion.retrieve}
\title{DepositoryVersion.retrieve}
\usage{
DepositoryVersion.retrieve(id)
}
\arguments{
\item{id}{String The ID or full name of a depository version}
}
\description{
Retrieves the metadata about a specific depository version.
}
\examples{
\dontrun{
DepositoryVersion.retrieve("ClinVar/1.0.0")
}

}
\references{
\url{https://docs.solvebio.com/}
}

