% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{Expression.evaluate}
\alias{Expression.evaluate}
\title{Expression.evaluate}
\usage{
Expression.evaluate(expression, data_type = "string", is_list = FALSE,
  data = NULL, raw = FALSE, env = solvebio:::.solveEnv)
}
\arguments{
\item{expression}{The SolveBio expression string.}

\item{data_type}{The data type to cast the expression result (default: string).}

\item{is_list}{Set to TRUE if the result is expected to be a list (default: FALSE).}

\item{data}{Variables used in the expression (default: NULL).}

\item{raw}{Set to TRUE to return the raw response (default: FALSE).}

\item{env}{(optional) Custom client environment.}
}
\description{
Evaluate a SolveBio expression.
}
\examples{
\dontrun{
Expression.evaluate("1 + 1", data_type="integer", is_list=FALSE)
}

}
\references{
\url{https://docs.solvebio.com/}
}
