% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_search.R
\name{GlobalSearch.facets}
\alias{GlobalSearch.facets}
\title{GlobalSearch.facets}
\usage{
GlobalSearch.facets(facets, env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{facets}{Facets list.}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters (e.g. filters, entities, query).}
}
\description{
Performs a Global Search based on provided filters, entities, queries, and returns an R data frame containing the facets results from API response.
}
\examples{
\dontrun{
GlobalSearch.facets(facets="study")
}

}
\references{
\url{https://docs.solvebio.com/}
}
