% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_search.R
\name{GlobalSearch.results}
\alias{GlobalSearch.results}
\title{GlobalSearch.results}
\usage{
GlobalSearch.results(
  paginate = FALSE,
  use_field_titles = TRUE,
  env = solvebio:::.solveEnv,
  ...
)
}
\arguments{
\item{paginate}{When set to TRUE, retrieves all records (memory permitting).}

\item{use_field_titles}{(optional) Use field title instead of field name for query.}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters (e.g. filters, entities, query, limit, offset).}
}
\description{
Performs a Global Search based on provided filters, entities, queries, and returns an R data frame containing results from API response.
Returns a single page of results otherwise (default).
}
\examples{
\dontrun{
# No filters applied
GlobalSearch.results()

#Global Beacon search
GlobalSearch.results(entities = '[["gene","BRCA2"]]')

GlobalSearch.results(filters = '[{"and":[["type__in",["vault"]]]}]')

# Advanced search
GlobalSearch.results(query = "fuji")
}

}
\references{
\url{https://docs.solvebio.com/}
}
