% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_search.R
\name{GlobalSearch.search}
\alias{GlobalSearch.search}
\title{GlobalSearch.search}
\usage{
GlobalSearch.search(filters, env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{filters}{(optional) Query filters.}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters (e.g. limit, offset).}
}
\description{
Performs Global Search based on provided filters.
}
\examples{
\dontrun{
# No filters are applied
GlobalSearch.search()

# Global Beacon search
GlobalSearch.search('{"entities":[["gene","BRCA2"]]}')

# Type filter (only vaults)
GlobalSearch.search('{"filters":[{"and":[["type__in",["vault"]]]}]}')

# Advanced search
GlobalSearch.search('{"filters":[{"and":[{"and":[["created_at__range",["2020-12-27","2021-12-27"]]]}]}],"query":"fuji"}')

}

}
\references{
\url{https://docs.solvebio.com/}
}
