% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_search.R
\name{GlobalSearch.subjects}
\alias{GlobalSearch.subjects}
\title{GlobalSearch.subjects}
\usage{
GlobalSearch.subjects(env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters (e.g. filters, entities, query, limit, offset).}
}
\description{
Performs a Global Search based on provided filters, entities, queries, and returns an R data frame containing subjects from API response.
}
\examples{
\dontrun{
GlobalSearch.subjects(entities = '[["gene","BRCA2"]]')
}

}
\references{
\url{https://docs.solvebio.com/}
}
