% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/client.R
\name{login}
\alias{login}
\title{login}
\usage{
login(api_key, api_host, envir = solvebio:::.solveEnv$current)
}
\arguments{
\item{api_key}{Your SolveBio API key}

\item{api_host}{SolveBio API host (default: https://api.solvebio.com)}
}
\description{
Store and verify your SolveBio credentials.
}
\examples{
\dontrun{
login()
}
}
\references{
\url{https://docs.solvebio.com/}
}

