% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.all}
\alias{Object.all}
\title{Object.all}
\usage{
Object.all(env = .solveEnv, ...)
}
\arguments{
\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters.}
}
\description{
Retrieves the metadata about all objects on SolveBio accessible to the current user.
}
\examples{
\dontrun{
Object.all()
}

}
\references{
\url{https://docs.solvebio.com/}
}
