% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.count}
\alias{Object.count}
\title{Object.count}
\usage{
Object.count(id, ...)
}
\arguments{
\item{id}{The ID of a SolveBio object, or a Object object.}

\item{...}{(optional) Additional query parameters (e.g. filters, limit, offset).}
}
\description{
Returns the total number of records for a given SolveBio object.
}
\examples{
\dontrun{
Object.count("ClinVar/Variants")
Object.count("ClinVar/Variants", filters='[["gene_symbol", "BRCA2"]]')
}

}
\references{
\url{https://docs.solvebio.com/}
}
