% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.create}
\alias{Object.create}
\title{Object.create}
\usage{
Object.create(depository_version_id, name, ...)
}
\arguments{
\item{depository_version_id}{The ID of the parent depository verison.}

\item{name}{The unique name of the object witin the version.}

\item{...}{(optional) Additional object attributes.}
}
\description{
Create an empty SolveBio object.
}
\examples{
\dontrun{
Object.create(depository_version_id=1, name="MyObject")
}

}
\references{
\url{https://docs.solvebio.com/}
}
