% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.data}
\alias{Object.data}
\title{Object.data}
\usage{
Object.data(id, filters, ...)
}
\arguments{
\item{id}{The ID of a SolveBio object, or a Object object.}

\item{filters}{(optional) Query filters.}

\item{...}{(optional) Additional query parameters (e.g. limit, offset).}
}
\description{
Returns one page of documents from a SolveBio object and processes the response.
}
\examples{
\dontrun{
Object.data("ClinVar/ClinVar")
}

}
\references{
\url{https://docs.solvebio.com/}
}
