% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.facets}
\alias{Object.facets}
\title{Object.facets}
\usage{
Object.facets(id, facets, ...)
}
\arguments{
\item{id}{The ID of a SolveBio object, or a Object object.}

\item{facets}{A list of one or more field facets.}

\item{...}{(optional) Additional query parameters (e.g. filters, limit, offset).}
}
\description{
Retrieves aggregated statistics or term counts for one or more fields in a SolveBio object. Returns a list of data frames, one for each requested facet.
}
\examples{
\dontrun{
Object.facets("ClinVar/Combined", list("clinical_significance", "gene_symbol"))
}

}
\references{
\url{https://docs.solvebio.com/}
}
