% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.get_by_full_path}
\alias{Object.get_by_full_path}
\title{Object.get_by_full_path}
\usage{
Object.get_by_full_path(full_path, env = .solveEnv, ...)
}
\arguments{
\item{full_path}{The full path to the object.}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters.}
}
\description{
A helper function to get an object on SolveBio by its full path.
}
\examples{
\dontrun{
Object.get_by_full_path("solvebio:public:/ClinVar")
}

}
\references{
\url{https://docs.solvebio.com/}
}
