% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.get_by_path}
\alias{Object.get_by_path}
\title{Object.get_by_path}
\usage{
Object.get_by_path(path, ...)
}
\arguments{
\item{path}{The path to the object, relative to a vault.}

\item{...}{(optional) Additional query parameters.}
}
\description{
A helper function to get an object on SolveBio by its path. Used as a pass-through function from some Vault methods.
}
\examples{
\dontrun{
Object.get_by_path("/ClinVar")
}

}
\references{
\url{https://docs.solvebio.com/}
}
