% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.get_download_url}
\alias{Object.get_download_url}
\title{Object.get_download_url}
\usage{
Object.get_download_url(id, env = .solveEnv)
}
\arguments{
\item{id}{The ID of the object.}

\item{env}{(optional) Custom client environment.}
}
\description{
Helper method to get the download URL for a file object.
}
\examples{
\dontrun{
Object.get_download_url("1234567890")
}

}
\references{
\url{https://docs.solvebio.com/}
}
