% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.get_global_beacon_status}
\alias{Object.get_global_beacon_status}
\title{Object.get_global_beacon_status}
\usage{
Object.get_global_beacon_status(id, env = solvebio:::.solveEnv)
}
\arguments{
\item{id}{The ID of a SolveBio dataset.}

\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the global beacon status for the specified dataset.
}
\examples{
\dontrun{
Object.get_global_beacon_status("1234567890")
}

}
\references{
\url{https://docs.solvebio.com/}
}
