% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.query}
\alias{Object.query}
\title{Object.query}
\usage{
Object.query(id, paginate = FALSE, ...)
}
\arguments{
\item{id}{The ID of a SolveBio object, or a Object object.}

\item{paginate}{When set to TRUE, retrieves up to 500,000 records.}

\item{...}{(optional) Additional query parameters (e.g. filters, limit, offset).}
}
\description{
Queries a SolveBio object and returns an R data frame containing all records (up to 500,000).
Returns a single page of results otherwise (default).
}
\examples{
\dontrun{
Object.query("ClinVar/ClinVar", paginate=TRUE)
}

}
\references{
\url{https://docs.solvebio.com/}
}
