% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.retrieve}
\alias{Object.retrieve}
\title{Object.retrieve}
\usage{
Object.retrieve(id)
}
\arguments{
\item{id}{String The ID of a SolveBio object}
}
\description{
Retrieves the metadata about a specific object from SolveBio.
}
\examples{
\dontrun{
Object.retrieve("ClinVar/ClinVar")
}

}
\references{
\url{https://docs.solvebio.com/}
}
