% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.update}
\alias{Object.update}
\title{Object.update}
\usage{
Object.update(id, env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{id}{The ID of the vault to update.}

\item{env}{(optional) Custom client environment.}

\item{...}{Object attributes to change.}
}
\description{
Updates the attributes of an existing vault object.
}
\examples{
\dontrun{
Object.update(
              id="1234",
              filename="New Name",
             )
}

}
\references{
\url{https://docs.solvebio.com/}
}
