% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{protectedServer}
\alias{protectedServer}
\title{protectedServer}
\usage{
protectedServer(server, client_id, base_url)
}
\arguments{
\item{server}{Your original Shiny server function.}

\item{client_id}{Your application's client ID.}

\item{base_url}{(optional) Override the default login host (default: https://my.solvebio.com).}
}
\description{
Wraps an existing Shiny server in an OAuth2 flow.
}
\examples{
\dontrun{
protectedServer(
                server=server,
                client_id="abcd1234"
                )
}

}
\references{
\url{https://docs.solvebio.com/}
}
