% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SolveBioClient.R
\name{SolveBioClient}
\alias{SolveBioClient}
\title{SolveBioClient}
\usage{
SolveBioClient(api_key = character(), api_host = character())
}
\arguments{
\item{api_key}{Your SolveBio API key}

\item{api_host}{SolveBio API host (default: https://api.solvebio.com)}
}
\description{
Create an SolveBio client instance
}
\examples{
client <- SolveBioClient()
client$isAuthenticated()
}
\references{
\url{https://docs.solvebio.com/}
}

