% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{Task.all}
\alias{Task.all}
\title{Task.all}
\usage{
Task.all(env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters.}
}
\description{
Retrieves the metadata about all tasks on SolveBio accessible to the current user.
}
\examples{
\dontrun{
Task.all()
}

}
\references{
\url{https://docs.solvebio.com/}
}
