% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{Task.follow}
\alias{Task.follow}
\title{Task.follow}
\usage{
Task.follow(id, env = solvebio:::.solveEnv)
}
\arguments{
\item{id}{String The ID of a task.}

\item{env}{(optional) Custom client environment.}
}
\description{
A helper function to follow a specific tasks.
}
\examples{
\dontrun{
Task.follow("1234567890")
}

}
\references{
\url{https://docs.solvebio.com/}
}
