% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{Task.retrieve}
\alias{Task.retrieve}
\title{Task.retrieve}
\usage{
Task.retrieve(id, env = solvebio:::.solveEnv)
}
\arguments{
\item{id}{The ID of a task.}

\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the metadata about a specific task on SolveBio.
}
\examples{
\dontrun{
Task.retrieve("1234567890")
}

}
\references{
\url{https://docs.solvebio.com/}
}
