% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{Upload.retrieve}
\alias{Upload.retrieve}
\title{Upload.retrieve}
\usage{
Upload.retrieve(id, env = solvebio:::.solveEnv)
}
\arguments{
\item{id}{The ID a SolveBio upload}

\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the metadata about a specific upload from SolveBio.
}
\examples{
\dontrun{
Upload.retrieve(123)
}

}
\references{
\url{https://docs.solvebio.com/}
}
