% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{User.retrieve}
\alias{User.retrieve}
\title{User.retrieve}
\usage{
User.retrieve(env = solvebio:::.solveEnv)
}
\arguments{
\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves information about the current user.
}
\examples{
\dontrun{
User.retrieve()
}

}
\references{
\url{https://docs.solvebio.com/}
}
