% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{Vault.all}
\alias{Vault.all}
\title{Vault.all}
\usage{
Vault.all(..., env = .solveEnv)
}
\arguments{
\item{...}{(optional) Additional query parameters (e.g. limit, offset).}

\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the metadata about all accessible vaults.
}
\examples{
\dontrun{
Vault.all()
}

}
\references{
\url{https://docs.solvebio.com/}
}
