% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{Vault.create_dataset}
\alias{Vault.create_dataset}
\title{Vault.create_dataset}
\usage{
Vault.create_dataset(id, path, name, env = .solveEnv, ...)
}
\arguments{
\item{id}{The ID of the vault.}

\item{path}{The path to the dataset, within the vault.}

\item{name}{The name (filename) for the dataset.}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional dataset creation parameters.}
}
\description{
Create a new dataset in a vault.
}
\examples{
\dontrun{
vault = Vault.get_personal_vault()
Vault.create_dataset(vault$id, path="/", name="My Dataset")
}

}
\references{
\url{https://docs.solvebio.com/}
}
