% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{Vault.create_folder}
\alias{Vault.create_folder}
\title{Vault.create_folder}
\usage{
Vault.create_folder(id, path, name, ...)
}
\arguments{
\item{id}{The ID of the vault.}

\item{path}{The path to the folder, within the vault.}

\item{name}{The name (filename) of the new folder.}

\item{...}{(optional) Additional folder creation parameters.}
}
\description{
Create a new folder in a vault.
}
\examples{
\dontrun{
vault = Vault.get_personal_vault()
Vault.create_folder(vault$id, path="/", name="My Folder")
}

}
\references{
\url{https://docs.solvebio.com/}
}
