% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{Vault.get_or_create_by_full_path}
\alias{Vault.get_or_create_by_full_path}
\title{Vault.get_or_create_by_full_path}
\usage{
Vault.get_or_create_by_full_path(full_path, env = solvebio:::.solveEnv,
  ...)
}
\arguments{
\item{full_path}{The full path of a SolveBio vault.}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional parameters.}
}
\description{
Retrieves or creates a specific vault by its full path (domain:vault).
}
\examples{
\dontrun{
Vault.get_or_create_by_full_path("My New Vault")
}

}
\references{
\url{https://docs.solvebio.com/}
}
