% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{Vault.get_personal_vault}
\alias{Vault.get_personal_vault}
\title{Vault.get_personal_vault}
\usage{
Vault.get_personal_vault(env = .solveEnv)
}
\arguments{
\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the current users"s personal, private vault.
}
\examples{
\dontrun{
Vault.get_personal_vault()
}

}
\references{
\url{https://docs.solvebio.com/}
}
