% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{Vault.objects}
\alias{Vault.objects}
\title{Vault.objects}
\usage{
Vault.objects(id, ...)
}
\arguments{
\item{id}{The ID of the vault.}

\item{...}{(optional) Additional query parameters (e.g. limit, offset).}
}
\description{
Retrieves all objects in a specific vault.
}
\examples{
\dontrun{
vault = Vault.get_personal_vault()
Vault.objects(vault$id)
}

}
\references{
\url{https://docs.solvebio.com/}
}
