% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{Vault.retrieve}
\alias{Vault.retrieve}
\title{Vault.retrieve}
\usage{
Vault.retrieve(id)
}
\arguments{
\item{id}{String The ID of a SolveBio vault}
}
\description{
Retrieves the metadata about a specific SolveBio vault.
}
\examples{
\dontrun{
Vault.retrieve("1234567890")
}

}
\references{
\url{https://docs.solvebio.com/}
}
