% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{Vault.search}
\alias{Vault.search}
\title{Vault.search}
\usage{
Vault.search(id, query, env = .solveEnv, ...)
}
\arguments{
\item{id}{The ID of the vault.}

\item{query}{The search query.}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters (e.g. limit, offset).}
}
\description{
Search for objects in a specific vault.
}
\examples{
\dontrun{
vault = Vault.get_personal_vault()
Vault.search('test')
}

}
\references{
\url{https://docs.solvebio.com/}
}
