% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beacon.R
\name{Beacon.all}
\alias{Beacon.all}
\title{Beacon.all}
\usage{
Beacon.all(env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters.}
}
\description{
Retrieves the metadata about all beacons on SolveBio accessible to the current user.
}
\examples{
\dontrun{
Beacon.all()
}

}
\references{
\url{https://docs.solvebio.com/}
}
