% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\name{Dataset.all}
\alias{Dataset.all}
\title{Dataset.all}
\usage{
Dataset.all(...)
}
\description{
Retrieves the metadata about all datasets on SolveBio.
}
\examples{
\dontrun{
Dataset.all()
}
}
\references{
\url{https://docs.solvebio.com/}
}

