% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset.all}
\alias{Dataset.all}
\title{Dataset.all}
\usage{
Dataset.all(env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters (e.g. page).}
}
\description{
Retrieves the metadata about datasets on SolveBio.
}
\examples{
\dontrun{
Dataset.all()
}

}
\references{
\url{https://docs.solvebio.com/}
}
