% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset.count}
\alias{Dataset.count}
\title{Dataset.count}
\usage{
Dataset.count(id, ...)
}
\arguments{
\item{id}{The ID or full name of a SolveBio dataset, or a Dataset object.}

\item{...}{(optional) Additional query parameters (e.g. filters, limit, offset).}
}
\description{
Returns the total number of records for a given SolveBio dataset.
}
\examples{
\dontrun{
login()
Dataset.count("ClinVar/Variants")
Dataset.count("ClinVar/Variants", filters='[["gene_symbol", "BRCA2"]]')
}

}
\references{
\url{https://docs.solvebio.com/}
}

