% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset.create}
\alias{Dataset.create}
\title{Dataset.create}
\usage{
Dataset.create(
  vault_id,
  vault_parent_object_id,
  name,
  env = solvebio:::.solveEnv,
  ...
)
}
\arguments{
\item{vault_id}{The ID of the vault.}

\item{vault_parent_object_id}{The parent object (folder) ID in the vault.}

\item{name}{The name of the dataset in the parent folder.}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional dataset attributes.}
}
\description{
Create an empty SolveBio dataset.
}
\examples{
\dontrun{
Dataset.create(vault_id=vault$id, vault_parent_object_id=NULL, name="My Dataset")
}

}
\references{
\url{https://docs.solvebio.com/}
}
