% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset.delete}
\alias{Dataset.delete}
\title{Dataset.delete}
\usage{
Dataset.delete(id, env = solvebio:::.solveEnv)
}
\arguments{
\item{id}{String The ID of a SolveBio dataset}

\item{env}{(optional) Custom client environment.}
}
\description{
Delete a specific dataset from SolveBio.
}
\examples{
\dontrun{
Dataset.delete("1234567890")
}

}
\references{
\url{https://docs.solvebio.com/}
}
