% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\name{Dataset.query}
\alias{Dataset.query}
\title{Dataset.query}
\usage{
Dataset.query(id, filters = NULL, params = list())
}
\arguments{
\item{id}{String The ID or full name of a SolveBio dataset}

\item{filters}{Json}

\item{params}{Json Additional query parameters}
}
\description{
Returns filtered documents from a SolveBio dataset.
\code{query} executes a SolveBio dataset query and retrieves the results.
}
\examples{
\dontrun{
login()
Dataset.query("ClinVar/ClinVar")
}
}
\references{
\url{https://docs.solvebio.com/}
}

