% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset.query}
\alias{Dataset.query}
\title{Dataset.query}
\usage{
Dataset.query(id, filters, ...)
}
\arguments{
\item{id}{The ID or full name of a SolveBio dataset, or a Dataset object.}

\item{filters}{(optional) Query filters.}

\item{...}{(optional) Additional query parameters (e.g. limit, offset).}
}
\description{
Returns filtered documents from a SolveBio dataset.
\code{query} executes a SolveBio dataset query and retrieves the results.
}
\examples{
\dontrun{
login()
Dataset.query("ClinVar/ClinVar")
}

}
\references{
\url{https://docs.solvebio.com/}
}

