% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\name{Dataset.retrieve}
\alias{Dataset.retrieve}
\title{Dataset.retrieve}
\usage{
Dataset.retrieve(id, ...)
}
\arguments{
\item{id}{String The ID or full name of a SolveBio dataset}
}
\description{
Retrieves the metadata about a specific dataset from SolveBio.
}
\examples{
\dontrun{
Dataset.retrieve("ClinVar/ClinVar")
}
}
\references{
\url{https://docs.solvebio.com/}
}

