% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset.retrieve}
\alias{Dataset.retrieve}
\title{Dataset.retrieve}
\usage{
Dataset.retrieve(id, env = .solveEnv)
}
\arguments{
\item{id}{String The ID of a SolveBio dataset}

\item{env}{(optional) Custom client environment.}
}
\description{
Retrieves the metadata about a specific dataset from SolveBio.
}
\examples{
\dontrun{
Dataset.retrieve("1234567890")
}

}
\references{
\url{https://docs.solvebio.com/}
}
