% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_export.R
\name{DatasetExport.delete}
\alias{DatasetExport.delete}
\title{DatasetExport.delete}
\usage{
DatasetExport.delete(id, env = .solveEnv)
}
\arguments{
\item{id}{String The ID of a SolveBio dataset export.}

\item{env}{(optional) Custom client environment.}
}
\description{
Deletes a specific dataset export on SolveBio.
}
\examples{
\dontrun{
DatasetExport.delete(<ID>)
}

}
\references{
\url{https://docs.solvebio.com/}
}
