% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_export.R
\name{DatasetExport.retrieve}
\alias{DatasetExport.retrieve}
\title{DatasetExport.retrieve}
\usage{
DatasetExport.retrieve(id)
}
\arguments{
\item{id}{String The ID or full name of a SolveBio dataset export.}
}
\description{
Retrieves the metadata about a specific dataset export on SolveBio.
}
\examples{
\dontrun{
DatasetExport.retrieve(<ID>)
}

}
\references{
\url{https://docs.solvebio.com/}
}
