% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_field.R
\name{DatasetField.create}
\alias{DatasetField.create}
\title{DatasetField.create}
\usage{
DatasetField.create(dataset_id, name, data_type = "auto", ...)
}
\arguments{
\item{dataset_id}{The dataset ID.}

\item{name}{The name of the dataset field.}

\item{data_type}{(optional) The data type for the field (default: auto).}

\item{...}{(optional) Additional dataset import attributes.}
}
\description{
Create a new dataset field.
}
\examples{
\dontrun{
DatasetField.create(dataset_id=<ID>, name="my_field", title="My Field", data_type="string")
}

}
\references{
\url{https://docs.solvebio.com/}
}

