% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_field.R
\name{DatasetField.retrieve}
\alias{DatasetField.retrieve}
\title{DatasetField.retrieve}
\usage{
DatasetField.retrieve(id)
}
\arguments{
\item{id}{String The ID or full name of a dataset field.}
}
\description{
Retrieves the metadata about a specific dataset field.
}
\examples{
\dontrun{
DatasetField.retrieve(691)
}

}
\references{
\url{https://docs.solvebio.com/}
}

