% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_import.R
\name{DatasetImport.create}
\alias{DatasetImport.create}
\title{DatasetImport.create}
\usage{
DatasetImport.create(dataset_id, upload_id, manifest = NULL,
  commit_mode = "append", auto_approve = TRUE, ...)
}
\arguments{
\item{dataset_id}{The target dataset ID.}

\item{upload_id}{An upload ID}

\item{manifest}{(optional) A valid file manifest (alternative to upload_id).}

\item{commit_mode}{(optional) The commit mode (default: append).}

\item{auto_approve}{(optional) Automatically approve the commit (default: TRUE).}

\item{...}{(optional) Additional dataset import attributes.}
}
\description{
Create a new dataset import.
}
\examples{
\dontrun{
DatasetImport.create(dataset_id=<ID>, upload_id=<ID>)
}

}
\references{
\url{https://docs.solvebio.com/}
}
