% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depository_version.R
\name{DepositoryVersion.all}
\alias{DepositoryVersion.all}
\title{DepositoryVersion.all}
\usage{
DepositoryVersion.all(...)
}
\arguments{
\item{...}{(optional) Additional query parameters (e.g. page).}
}
\description{
Retrieves the metadata about all depository versions on SolveBio.
}
\examples{
\dontrun{
DepositoryVersion.all()
}

}
\references{
\url{https://docs.solvebio.com/}
}

