% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.create}
\alias{Object.create}
\title{Object.create}
\usage{
Object.create(vault_id, parent_object_id, object_type, filename, ...)
}
\arguments{
\item{vault_id}{The target vault ID.}

\item{parent_object_id}{The ID of the parent object (folder) or NULL for the vault root.}

\item{object_type}{The type of object (i.e. "folder").}

\item{filename}{The filename (i.e. the name) of the object.}

\item{...}{(optional) Additional object attributes.}
}
\description{
Create a SolveBio object.
}
\examples{
\dontrun{
Object.create(
              vault_id="1234567890",
              parent_object_id=NULL,
              object_type="folder",
              filename="My Folder"
              )
}

}
\references{
\url{https://docs.solvebio.com/}
}
