% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.delete}
\alias{Object.delete}
\title{Object.delete}
\usage{
Object.delete(id)
}
\arguments{
\item{id}{The ID of the object.}
}
\description{
Delete a specific object from SolveBio.
}
\examples{
\dontrun{
Object.delete("1234567890")
}

}
\references{
\url{https://docs.solvebio.com/}
}
