% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.retrieve}
\alias{Object.retrieve}
\title{Object.retrieve}
\usage{
Object.retrieve(id)
}
\arguments{
\item{id}{The ID of the object.}
}
\description{
Retrieves the metadata about a specific object on SolveBio.
}
\examples{
\dontrun{
Object.retrieve("1234567890")
}

}
\references{
\url{https://docs.solvebio.com/}
}
