% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Object.upload_file}
\alias{Object.upload_file}
\title{Object.upload_file}
\usage{
Object.upload_file(local_path, vault_id, vault_path, filename)
}
\arguments{
\item{local_path}{The path to the local file}

\item{vault_id}{The SolveBio vault ID}

\item{vault_path}{The remote path in the vault}

\item{filename}{(optional) The filename for the uploaded file in the vault (default: the basename of the local_path)}
}
\description{
Upload a local file to a vault on SolveBio. The vault path provided is the parent directory for uploaded file.
}
\examples{
\dontrun{
Object.upload_file("my_file.json.gz", vault$id, "/path/to/my_file.json.gz")
}

}
\references{
\url{https://docs.solvebio.com/}
}
