% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{protectedServerJS}
\alias{protectedServerJS}
\title{protectedServerUI}
\usage{
protectedServerJS()
}
\description{
Returns ShinyJS-compatible JS code to support cookie-based token storage.
}
\examples{
\dontrun{
jscookie_src <- "https://cdnjs.cloudflare.com/ajax/libs/js-cookie/2.2.0/js.cookie.js"
ui <- fluidPage(
    shiny::tags$head(
        shiny::tags$script(src = jscookie_src)
    ),
    useShinyjs(),
    extendShinyjs(text = solvebio::protectedServerJS(),
                  functions = c("enableCookieAuth", "getCookie", "setCookie", "rmCookie"))
)
}

}
\references{
\url{https://docs.solvebio.com/}
}
