% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{query.solvebio}
\title{query}
\usage{
query(client, dataset, filters, params)

\method{query}{solvebio}(client, dataset, filters = NULL, params = NULL)
}
\arguments{
\item{client}{SolveBioClient A SolveBioClient object}

\item{dataset}{String The ID or full name of a SolveBio dataset}

\item{filters}{Json}

\item{params}{Json Additional query parameters}
}
\description{
Returns filtered documents from a SolveBio dataset.
\code{query} executes a SolveBio dataset query and retrieves the results.
}
\examples{
sb = SolveBioClient(Sys.getenv("SOLVEBIO_API_KEY"))
query(sb, dataset="ClinVar/ClinVar")
}
\references{
\url{https://docs.solvebio.com/}
}

