% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/client.R
\name{SolveBioClient}
\alias{SolveBioClient}
\title{SolveBioClient}
\usage{
SolveBioClient(api_key, api_host = "https://api.solvebio.com")
}
\arguments{
\item{api_key}{Your SolveBio API key}

\item{api_host}{SolveBio API host (default: https://api.solvebio.com)}
}
\description{
Create an SolveBio client instance
}
\examples{
SolveBioClient(Sys.getenv("SOLVEBIO_API_KEY"))
}
\references{
\url{https://docs.solvebio.com/}
}

