% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{Task.follow}
\alias{Task.follow}
\title{Task.follow}
\usage{
Task.follow(id, env = .solveEnv, interval = 2)
}
\arguments{
\item{id}{String The ID of a task.}

\item{env}{(optional) Custom client environment.}

\item{interval}{2. Delay in seconds between each completion status query}
}
\value{
the task object
}
\description{
A helper function to follow a specific task until it gets completed.
}
\examples{
\dontrun{
Task.follow("1234567890")
}
}
\references{
\url{https://docs.solvebio.com/}
}
