% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{Upload.all}
\alias{Upload.all}
\title{Upload.all}
\usage{
Upload.all(...)
}
\arguments{
\item{...}{(optional) Additional query parameters (e.g. page).}
}
\description{
Retrieves the metadata about all uploads on SolveBio.
}
\examples{
\dontrun{
Upload.all()
}

}
\references{
\url{https://docs.solvebio.com/}
}

