% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{Upload.create}
\alias{Upload.create}
\title{Upload.create}
\usage{
Upload.create(path)
}
\arguments{
\item{path}{The path to the local file}
}
\description{
Upload a local file to SolveBio.
}
\examples{
\dontrun{
Upload.create("my_file.json.gz")
}

}
\references{
\url{https://docs.solvebio.com/}
}
