% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{Vault.create}
\alias{Vault.create}
\title{Vault.create}
\usage{
Vault.create(name, env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{name}{The unique name of the vault.}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional vault attributes.}
}
\description{
Create a new SolveBio vault.
}
\examples{
\dontrun{
Vault.create(name="my-domain:MyVault")
}

}
\references{
\url{https://docs.solvebio.com/}
}
