% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{Vault.create_folder}
\alias{Vault.create_folder}
\title{Vault.create_folder}
\usage{
Vault.create_folder(id, path, recursive = FALSE,
  env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{id}{The ID of the vault.}

\item{path}{The path to the folder, within the vault.}

\item{recursive}{Create all parent directories that do not yet exist (default: FALSE).}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional folder creation parameters.}
}
\description{
Create a new folder in a vault.
}
\examples{
\dontrun{
vault = Vault.get_personal_vault()
Vault.create_folder(vault$id, "/My Folder")
}

}
\references{
\url{https://docs.solvebio.com/}
}
