% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{Vault.datasets}
\alias{Vault.datasets}
\title{Vault.datasets}
\usage{
Vault.datasets(id, env = solvebio:::.solveEnv, ...)
}
\arguments{
\item{id}{The ID of the vault.}

\item{env}{(optional) Custom client environment.}

\item{...}{(optional) Additional query parameters (e.g. limit, offset).}
}
\description{
Retrieves all datasets in a specific vault.
}
\examples{
\dontrun{
vault = Vault.get_personal_vault()
Vault.datasets(vault$id)
}

}
\references{
\url{https://docs.solvebio.com/}
}
